Fixed SVGF, faster Fog encoding, cleaned code

+++ shaders/include/internal_settings.glsl
@@ -25,7 +25,7 @@
 #endif
 
 /*
-const int  shadowcolor0Format      = RGBA8;
+const int  shadowcolor0Format      = RGBA16F;
 const bool shadowHardwareFiltering = false;
 
 const int  colortex0Format         = RGBA16;            // Main
